feat(write): remove `StrConsumer` and wrap `fmt`
